+++ packages/winnow/CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [1.5.1] - 02-10-2017
 ### Fixed
 * OutSR parameter handles latestWkid and wkt
 
@@ -75,6 +75,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 * Initial release
 
+[1.5.1]: https://github.com/dmfenton/winnow/compare/v1.5.0..v1.5.1
 [1.5.0]: https://github.com/dmfenton/winnow/compare/v1.4.2..v1.5.0
 [1.4.2]: https://github.com/dmfenton/winnow/compare/v1.4.1..v1.4.2
 [1.4.1]: https://github.com/dmfenton/winnow/compare/v1.4.0..v1.4.1